--- Animations/gitMergeEg.py
@@ -1,12 +1,0 @@
-from rich.console import Console
-import time
-
-# Initialize the console
-console = Console()
-
-def show_add_animation():
-    
-    
-
-
-show_add_animation()
